ldp: split LDP messages into multiple PDUs when necessary
The nbr_write_loop() function uses a timer to group as many messages
as possible within a single LDP PDU. However, the encoded PDU must not
surpass the maximum PDU length agreed upon in the session negotiation,
otherwise the receiving peer will be unable to decode the packet. To
address this issue, the Pdu::encode() function has been updated
to split messages into multiple LDP PDUs when necessary. This is a
generic solution that applies to all message types.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>