daemon: delay activation of confirmed commit timeout
To prevent the confirmed commit timeout from starting prematurely
in case the commit is rejected, it is now triggered only after a
successful commit.

Additionally, the code has been refactored to use the `Rollback` struct,
which ensures that the rollback configuration and timeout are always
either present or not present simultaneously, simplifying the code.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>