daemon-gnmi: update TLS configuration
Add the "enabled" option instead of implicitly enabling TLS when the
"[plugins.gnmi.tls]" section is present in the configuration. This
gives more flexibility to the user.

Additionally, the GnmiTls config struct has been renamed to Tls,
as it will be reused by the gRPC module in the upcoming commit.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>