ospfv3: update interpretation of configured instance ID values
In the last draft of the OSPF YANG module, the "instance-id" leaf had a
`"range "0 .. 31"` constraint, implying it was a relative instance ID,
and the actual instance ID had to be calculated based on RFC5838's
address-family instance ID ranges.

In the RFC update, the range was removed, so let's now interpret the
configured instance IDs as absolute values.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>