ospf: improve checks for MaxAge LSAs in the unit tests
In the unit tests, the age of the LSAs being flooded is omitted from
the reference outputs since it's not relevant (and not deterministic
as well). It's important, however, to show when the LSA has reached
MaxAge as it holds special significance.

It's worth noting that in the northbound operational data reference
outputs, LSAs that have reached MaxAge are already appropriately
marked with the "maxage" leaf.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>