ldp: implement the TCP MD5 signature option
There's nothing worse than unprotected LDP sessions. This commit
addresses that issue :)

The implementation is a bit tricky due to the fact that LDP neighbors
are discovered dynamically instead of being configured explicitly.
That means the configured passwords must be set in both the TCP
listening socket and connection sockets (when playing the active role
in the initialization FSM).

Support for TCP-AO (TCP Authentication Option) with stronger
authentication algorithms is planned for the future.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>