ripv2: implement cryptographic authentication support
This commit introduces support for cryptographic authentication in
accordance with RFC 4822 ("RIPv2 Cryptographic Authentication").

The implementation is currently limited to static keys and the MD5
algorithm. Support for keychains and the HMAC-SHA1 family of algorithms
is planned for the future.

Unit tests have been added to ensure the functionality of this new
feature.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>