ospf: create separate network tx/rx tasks per interface
This commit lays the groundwork for the upcoming implementation of
cryptographic authentication support.

The goal is to create separate network transmit and receive tasks for
each interface. By doing so, each task can be associated with different
authentication data, enabling authentication to be performed during
the encoding and decoding of packets. This is ideal since we want to
offload as much work as possible from the main OSPF task.

The SO_BINDTODEVICE sockoption is used to allow multiple raw sockets
sockets bound to the same address (but bound to different interfaces).

The unit tests data has been updated to accommodate an IPC message
modification.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>