ripv2: adjust maximum number of RTEs when authentication is enabled
According to RFC 2453, a Response packet in RIPv2 can contain up
to 25 RTEs. However, when cryptographic authentication is used,
the authentication trailer is encoded as a regular RTE. In that
case, the effective maximum number of RTEs should be 23, not 24,
to accommodate both the authentication header and trailer.

It's worth noting that the authentication trailer may exceed 20 bytes
when using algorithms other than Keyed-MD5. That fact should be taken
into consideration when implementing those algorithms (e.g. HMAC-SHA1).

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>