ospfv2: implement cryptographic authentication support
This commit introduces support for cryptographic authentication in
accordance with Appendix D of RFC 2328.

The implementation is currently limited to static keys, with support
for keychains planned for the future.

For performance reasons, authentication is performed in the network
Tx/Rx tasks while packets are encoded and decoded. However, the
check for non-decreasing sequence numbers is performed within the
OSPF instance task, which has access to the necessary neighbor data
structures.

The code has been structured to facilitate future implementation of
other authentication algorithms as specified in RFC 5709, RFC 7166
and RFC 7474.

Unit tests have been included to verify the basic functionality.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>