ospfv2: fix authentication verification with LLS data blocks
Found an interoperability bug with IOS-XR which caused holo to
reject valid Hello packets that contained an LLS block, but only
when authentication was enabled. The issue stemmed from an incorrect
assumption regarding the extent of the authentication trailer, assuming
it extended until the end of the packet. This commit addresses the
problem by using the "Auth Data Len" field from the OSPF header to
correctly retrieve the authentication trailer. Now holo will ignore
LLS data blocks entirely as it should.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>