ospf: refactor interface for initializing the Options field
In OSPF, the same Options field is present in both OSPF Hello and
Database Description packets, as well as in specific LSA types.
The generic Area::area_options() function is used to return the
appropriate Options bits based on the area settings and where the
Options field will be included (packet or LSA).

To accommodate future requirements, where additional variables will be
required to determine the correct Options bits (e.g., authentication
settings affecting the OSPFv3 AT-bit), this commit introduces the
OptionsLocation enum. This enum provides a higher-level interface to
Area::area_options(), eliminating the need for multiple parameters
in the future.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>