ospf: introduce separate function for checksum offloading control
The OSPF protocol uses checksums to ensure packet integrity. In
OSPFv3, checksumming is more complex due to the inclusion of the IPv6
pseudo-header that is taken into account. For that reason, we use
the IPV6_CHECKSUM sockoption to offload OSPFv3 packet checksumming
to the Linux kernel.

However, when cryptographic authentication is enabled, packet
checksumming becomes redundant and should be disabled, as per RFC
2328 and RFC 7166.

In preparation for the upcoming OSPFv3 Authentication Trailer feature,
this commit separates the setting of the IPV6_CHECKSUM sockoption from
socket creation. This separation will allow us to enable or disable
checksum offloading as needed, based on the interface authentication
settings.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>