ospf: store boot count in non-volatile memory
The boot count represents the number of times an OSPF instance,
identified by its name, has been started. This count increments either
when the OSPF instance is individually restarted or when the router
undergoes a complete reboot.

The OSPF boot count will be used by the upcoming implementations of
RFC 7166 and RFC 7474 to ensure the strictly increasing property of
authentication sequence numbers. It may also have other potential
uses in the future.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>