daemon, protocols: give protocols access to non-volatile storage
Certain protocol functionalities, such as graceful restart and
cryptographic authentication, rely on the usage of non-volatile
storage for proper operation.

Currently, Holo uses non-volatile storage solely for configuration
transactions. This commit refactors the code to provide all protocol
crates with access to the non-volatile storage, which is loaded or
initialized by holod.

The non-volatile storage is implemented as a key-value store and is
now protected by a mutex to enable shared access. The non-volatile
storage is wrapped inside an Option type since it won't be available
in unit tests.

The "rollback-log" feature was removed and is now an integral part
of holod.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>