ospf: add support for authentication keychains
This commit implements the "key-chain" feature from the ietf-ospf
YANG module. Now, interfaces can be configured with authentication
keychains, providing smooth key rollover and improved security.

Under the hood, the AuthCtx structure was split into two separate
structs: AuthEncodeCtx and AuthDecodeCtx. The decision to do this
separation is that packet encoding and decoding require different
input parameters.

Packet encoding requires a single authentication key and access to the
global instance authentication sequence number. Also, by requiring the
key loopkup to be performed in advance, we can preserve the existing
API where packet encoding is infallible. Packet decoding, on the
other hand, requires access to the full interface keychain since the
key lookup depends on the Key-ID field of the received packets.

Additionally, the "src" parameter passed to the packet encode/decode
functions was moved to the corresponding AuthEncodeCtx and
AuthDecodeCtx structs, since the packet source address is only relevant
when authentication is enabled.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>