northbound: improve iteration over operational data
Extend the code responsible for iterating over YANG modules and
retrieving state data to efficiently handle single list entries.

As a side note, this part of the code is in need of a major refactoring
since there are lots of improvements and optimizations that can be
made. But let's leave that for another moment.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>