cli:  ‘address’ option to specify the holod address
A new argument "address" was added for holo-cli to give the ability to
the user to choose the address of holo-daemon which can be on the local
or remote machine with either of IPv4 or IPv6 address.
Using this feature you can give the exact address of holod.
The following formats are acceptable:

  - holo-cli --address http://127.0.0.1:50051
  - holo-cli -a http://127.0.0.1:50051
  - cargo run -- --address http://127.0.0.1:50051

Signed-off-by: Rasoul Mesghali <rasoul.mesghali@gmail.com>