protocols: transition to new netlink-based southbound infrastructure
This commit converts all protocol crates to use the new netlink-based
southbound infrastructure, based on ibus message passing with the
`holo-interface` and `holo-routing` base components. This is naturally
a big conversion. Also, the `holo-southbound` crate was removed, since
the zebra-based southbound infrastructure is no longer needed (yay!).

The conformance tests data will be updated in the following commit.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>