misc: increase internal bus message capacity
The current internal bus implementation uses a multi-producer,
multi-consumer broadcast channel from Tokio. The issue with broadcast
channels is that they drop old messages when the channel reaches its
full capacity, which is certainly undesirable.

In the long term, a refactor is needed to transition the internal bus
to a non-lossy channel with advanced features like pub-sub. However,
until that overhaul is complete, this commit addresses the immediate
problem by increasing the message capacity of the internal bus
to 1024. This change ensures that message dropping won't occur in
practical scenarios.

For details on the future plans regarding the internal bus, please
refer to this link:
https://github.com/users/rwestphal/projects/3?pane=issue&itemId=37067463

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>