yang: optimize generation of YANG textual representations
Most of the time, the textual representation of YANG values consists
of static strings. To avoid unnecessary memory allocations in these
cases, change `ToYang::to_yang` to return the `Cow` type, which can
accommodate both owned and static strings.

The northbound infrastructure for operational data still expects an
owned string for each state leaf, but that should change as part of
a large refactoring in the future.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>