bgp: refactor data structures to reduce memory consumption
In BGP, BTreeMaps are used to store prefixes and Adj-RIB routes. Due
to BTreeMaps overallocating nodes for performance reasons, and the
large volumes of data generally handled by BGP, it's important to
minimize the size of BTreeMap nodes. This is achieved by boxing
some fields, allocating them separately in the heap. This commit
reduces overall memory consumption by approximately 60%.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>