bgp: use boxed slice for storing unknown attributes
Once a list of unknown attributes is decoded from a message, we
don't need to manipulate that list later in terms of adding or
removing elements. In that case, we can switch to a boxed slice
since they consume less stack size compared to vectors (16 bytes
vs 24 bytes, respectively, which means 8 less bytes per route).

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>