bgp, routing: convert RIB to use prefix-trie
For storing routing tables, we need a data structure where we can
do longest-prefix matches, a requirement for things like BGP nexthop
resolution, policy application, among others.

For that, we'll be using the prefix-trie crate, which has excellent
performance numbers, an ergonomic API similar to std::collections,
and no additional dependencies.

The transition to prefix-trie is straightforward due to its API's
similarity to BTreeMap.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>