utils: implement `TryFromYang` for communities and large communities
These changes pave the way for future handling of route policy
configurations involving these community types.

For now, I haven't implemented the same trait for extended communities
and IPv6 extended communities. The YANG regular expressions for
these are known to be buggy [1], and how those community types are
represented is likely to change in future versions of the BGP YANG
model draft.

[1]
https://mailarchive.ietf.org/arch/msg/idr/mABe9mQgfktOLBPJBeRcOdeepQU/

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>