bgp: add initial test topologies
Add two basic BGP test topologies, including iBGP and eBGP sessions,
route redistribution and ECMP. These topologies should serve as a
foundation for upcoming conformance tests.

In a first moment, modify holo-bgp to skip relaying `PolicyApply`
messages to the test framework. This is necessary because the `nbr_tx`
task is also relaying messages to the test framework, so the order in
which those messages are collected is non-deterministic. A proper fix
will be implemented later.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>