northbound: cache YANG schema node data paths
Take advantage of the private pointer present in libyang's schema
nodes to cache the node's data path. This eliminates the need to
dynamically generate data paths whenever necessary, which is an
expensive operation. With this change, fetching operational data is
15% faster.

Since manipulating the schema node private pointers involves crossing
an FFI boundary, this commit makes use of two unsafe function calls.

Long term, the northbound framework should be refactored to rely almost
exclusively on auto-generated code from the YANG models. At that point,
this optimization should be removed as it will no longer be necessary.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>