northbound: remove dependency on unstable compiler feature
The latest nightly compiler was complaining that the
`unsized_fn_params` feature is internal to the compiler or standard
library and that using it is strongly discouraged.

In fact, `unsized_fn_params` is rather obscure and unlikely to be
stabilized anytime soon, if ever. As such, refactor the code such
that we no longer depend on it.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>