vrrp: handling for when VRRP priority is equal
When a vrrp packet is received, one of the things
we need to check is the source address for the IP
packet (as specified in the RFC 3768 pg 18 section
6.4.3), for handling one of the edge cases where
the priority is equal.

This had not yet been handled since we were not
fetching the sender IP address yet. Now that
`read_loop` handles that, we can handle this.

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>