vrrp: read_loop for ARP, cleaning up.
introduce the arp_rx task and `read_loop` for
ARP. This pipeline will be used when receiving
ARP packets externally.

Having the socket created inside the read_loop
instead of in another function is since the
sockets were not reading any ARP packets when
created normally. The only way the packets were
being read was through the actual `fd` which
the `as_raw_fd()` function was not providing.
`into_raw_fd()` was providing it, but caused the
fd of the socket to be moved which rendered it
useless whenever else it is used. We therefore
created a separate socket inside the `read_loop`
function until there is a better alternative.

The created socket has also been created
through libc and not Socket, for the same
reason.

remove run.sh which is only used locally.

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>