vrrp: introduce macvlan function
this commit introduces the function inside
the netlink.rs that will be used to create
the macvlan interface. This will later be
connected to the vrrp module where each
instance, if macvlan is turned on, should
use its own macvlan interface.

also added a line to `bgp.rs` in `holo-util`,
clippy had issue with very long first doc
paragraph.
[https://rust-lang.github.io/rust-clippy/master/index.html#/too_long_first_doc_paragraph]

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>