vrrp: initialize MacVlan on holo-vrrp module
On the holo-vrrp module, we added the
MacVlanInterface struct. This will be linked to
the respective InstanceCfgs, if the macvlan is
enabled for VRRP. The struct looks almost similar
to the Interface struct in holo-vrrp, but we
remove the vrrp instances and the tasks fields.

We also remove some unnecessary files.

Next should be to add calls to create these
macvlans inside vrrp then edit the VRRP fsm to
accomodate either the configuration of the
macvlan or not.

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>