vrrp: add create prepare. change mvlan name
Changed the name of the macvlan interface being
created to `mvlan-vrrp-{vrid}`. At the moment,
when creating the interfae, the messages are not
correctly received, so this simplifies the type
of interface we deal with.
Also add create_prepare in the configuration of
holo-interface to make sure the interface object
is correctly created.

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>