vrrp: deal with non-existent IP address
When the IP address did not exist in the parent
interface before, we still went ahead and
attempted to create the networks (which failed).

New modifications, there is a check of whenever
we have an IP address and an ifindex id
we want to add macvlan network information before
trying out any network related activity attached
to it.

The verification is also done during deletion of
an interface Address, where during deletion
 if the parent, if interface is left with no IP
 address, we will automatically remove all the
 network information of all instances attached to
 the Iterface.

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>