utils: expose multicast helpers for raw sockets
We shouldn't limit multicast setsockopt helpers to UDP sockets since
it's also possible to do multicasting using raw sockets (as in the
case of the upcoming VRRP implementation).

The methods moved from the `UdpSocketExt` trait to the `SocketExt`
trait were renamed to avoid conflicts with the native methods in the
socket2's `Socket` struct.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>