isis: fix logic for detecting LSP content changes
The correct behavior is to assume any installed LSP contains content
changes unless a previous instance exists in the LSPDB with identical
flags and TLVs. The existing code incorrectly treated LSPs without
a prior instance in the LSPDB as unchanged, which is wrong.

Properly detecting LSP content changes is necessary to avoid triggering
SPF unnecessarily when only the sequence number has changed.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>