isis: implement the SPF algorithm and route computation
This commit implements the following:
* SPF computation (both full runs and partial runs)
* Route computation for both IPv4 and IPv6 address families
* Installation of IS-IS routes in the global RIB
* Support for both old-style (narrow) and wide metrics

Of particular note, this implementation deviates slightly from the
ISO standard. Specifically, we don't pre-load TENT with the local
adjacency database. Instead, we pre-load from the local LSP. This was
done to allow SPF to be run with any router as the root, which will be
necessary in the future for the implementation of the TI-LFA feature.

Next steps:
* Conformance tests
* Implement multi-topology support (RFC 5120)
* Implement domain-wide prefix distribution (RFC 5302)

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>