bgp: validate message length in unit tests
In normal operation, `Message::get_message_len()` is used to
ensure the input buffer contains a complete BGP message before
decoding. Apply the same validation in unit tests to avoid passing
malformed data to the decoding function, which assumes the message
length is valid.

Additionally, simplify the logic for determining the presence of
the NLRI field to improve readability.

Fixes #44

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>