isis: limit SNP PDUs to the maximum LSP size
This change ensures compliance with the ISO specification, which
dictates that the size of SNP PDUs is constrained by the maximum LSP
size (and not by the interface MTU).

In scenarios where there's an MTU mismatch and Hello padding isn't
enabled, this change can help IS-IS function correctly on the link.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>