isis: fix sending of CSNP PDUs on large networks
The maximum number of LSP entries that fit in a single CSNP PDU
is 90, based on the default maximum LSP size of 1492. However,
`SnpTlvs::max_lsp_entries` was incorrectly calculating it as 91,
leading to CSNPs that were too large to be transmitted over the
network. This commit fixes an internal constant to resolve the problem.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>