northbound: optimize state retrieval
Extend the `ListEntryKind::child_task` method with a `module_name`
parameter. This addition allows the northbound iteration code to more
accurately identify which child tasks correspond to specific subtrees
of the YANG data hierarchy. As a result, unnecessary requests to child
tasks for non-relevant portions of the YANG data hierarchy are avoided.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>