vrrp3: merging of mvlan6 and mvlan4
mvlan6 & mvlan4 were previously used in
case the same vrid had a vrrp2 and vrrp3 instance. Since the
distinction of whether it is ipv4 or ipv6 instance is made
on the `Interface` level, there is no need to have them as
separate fields and both have been merged into a single
`mvlan` field.

Signed-off-by: Paul Wekesa <paul1tw1@gmail.com>