isis: change handling of interface southbound messages
Process interface system updates even when the IS-IS instance is
disabled. This ensures that interface information is available and
up-to-date when the IS-IS instance is re-enabled, eliminating the
need to re-request system information.

It should be noted that we only subscribe to updates for interfaces
configured under the IS-IS instance.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>