isis: display SPF back-off delay state data
Since SPF is scheduled separately for L1 and L2, add an augmentation
to display the SPF back-off delay state data for each level separately.

The IETF model assumes that SPF scheduling for L1 and L2 is handled
together, which is an oversight that will likely be addressed in
the future.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>