isis: add basic support for the Router Capability TLV
This adds encoding/decoding for the Router Capability TLV, plus some
supporting code. We're not including this TLV in LSPs yet, and it's
not used for anything at the moment.

That'll change soon with the Segment Routing implementation, where
we'll start using this TLV and add support for the related SR sub-TLVs.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>