isis: implement the SR Router Capability sub-TLVs
This is the first part of the SR-MPLS implementation. In addition
to supporting the SR Router Capability sub-TLVs, we now process the
local SR configuration and add the Router Capability TLV to the local
LSP with the appropriate sub-TLVs (SRGB and SRLB).

Next steps:
* Implement Prefix-SIDs
* Implement Adj-SIDs

As a side note, the ietf-isis-sr-mpls YANG module we're using was
slightly modified, as the original version has an augmentation issue.
This issue will be reported and should be addressed in the next
draft version.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>