routing: ignore "no such process" error when uninstalling a route
When an interface is shut down, all routes using that interface as
a nexthop are automatically removed by the Linux kernel. Routing
protocols also respond to the interface down event, such as by
rerunning SPF, and attempt to uninstall the affected routes if no
alternate path is available. In such cases, the attempt to uninstall
the route fails because the kernel has already removed it. As such,
this should be treated as a benign error and not logged.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>